:computer: Lots of changes
Not mostly UI changes, just rushed things for the deadline lol

Enemy now has separate collision box which the bullet hits and destroys the enemy.

Update is being used in the player to detect when their health is lower than 1 to end the game.

Added enemy generator script to the trench piece to generate enemies in the trench.

Player HUD has been added for ammo amount and health.

Bullet prefab has been modified to change the size as it wasn't detecting collisions correctly also changed to continuous detection in the RigidBody instead of discrete.

Added ExitGame method to game manager to go back to main menu when the player either dies or collides with the box in the end trench piece

Added collision box to end trench piece to take the player back to the main menu

Limited frame rate for the application to reduce component usage.

Also added a new camera layermask to stop jittering when shooting inside of the trenches